% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{The country of interest as a character string.}
}
\value{
A line graph of the population with respect to time.
}
\description{
Produces a line graph of the population growth from 1950 to 2020 of a
user specified country.
}
