% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup Attendance}
\format{
a data frame with 22 rows and 4 columns.
\describe{
  \item{WorldCup}{A unique identifier that gives the host country and year.}
  \item{Matches}{Total number of matches in that world cup.}
  \item{Totalattendance}{Total number of people who attended the matches.}
  \item{Averageattendance}{Average number of people per game.}
}
}
\source{
\url{'https://en.wikipedia.org/wiki/FIFA_World_Cup'}
}
\usage{
World_Cup
}
\description{
Gives the total attendance and average attendance for all world cups.
}
\keyword{datasets}
