% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population}
\format{
a data frame with 235 rows and 72 Columns
\describe{
  \item{Country}{The names of the country as a character string.}
  \item{1950}{Population in thousands in 1950 as a numeric.}
  \item{1951}{Population in thousands in 1951 as a numeric.}
  \item{etc}
}
}
\source{
{United Nations}
}
\usage{
WorldPopulation
}
\description{
Gives the population in thousands from 1950 t0 2020 for 235 countries
}
\keyword{datasets}
